Rename event bus module
