+++ examples/modules/events/event-bus/deployspec.yaml
