+++ examples/modules/events/event-bus/docs/_static/architecture.drawio
