+++ examples/modules/events/event-bus/requirements.txt
